ajuste no campo 'Valor' em lancamentos, para o tipo 'Decimal'

+++ SysContabil/src/Dominio/Dominio/Entidades/Lancamento.cs
@@ -6,7 +6,7 @@ namespace Dominio.Entidades
 {
     public class Lancamento
     {
-        public Lancamento(DateTime data, string debito, string credito, double valor, string reciboFiscal)
+        public Lancamento(DateTime data, string debito, string credito, decimal valor, string reciboFiscal)
         {
             Data = data;
             Debito = debito;
@@ -19,13 +19,13 @@ namespace Dominio.Entidades
         public DateTime Data { get; private set; }
         public string Debito { get; private set; }
         public string Credito { get; private set; }
-        public double Valor { get; private set; }
+        public decimal Valor { get; private set; }
         public string ReciboFiscal { get; private set; }
 
         public string PlanoDeContaID { get; private set; }
         public PlanoDeConta PlanoDeConta { get; private set; }
 
-        public void AtualizarLancamento(DateTime data, string debito, string credito, double valor, string reciboFiscal)
+        public void AtualizarLancamento(DateTime data, string debito, string credito, decimal valor, string reciboFiscal)
         {
             this.Data = data;
             this.Debito = debito;
